LPS-179214 Extract ScreenNavigationEntry methods from AccountGroupDetailsScreenNavigationCategory to AccountGroupDetailsScreenNavigationEntry, extends AccountGroupDetailsScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
