+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountGroupDetailsScreenNavigationCategory.java
@@ -15,25 +15,10 @@
 package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
-import com.liferay.account.admin.web.internal.display.AccountGroupDisplay;
-import com.liferay.account.admin.web.internal.security.permission.resource.AccountGroupPermission;
-import com.liferay.account.constants.AccountActionKeys;
-import com.liferay.account.constants.AccountConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-import com.liferay.portal.kernel.service.permission.PortalPermission;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -42,15 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Albert Lee
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class AccountGroupDetailsScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<AccountGroupDisplay> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -58,13 +39,8 @@ public class AccountGroupDetailsScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_DETAILS;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, "details");
+		return language.get(locale, "details");
 	}
 
 	@Override
@@ -73,41 +49,7 @@ public class AccountGroupDetailsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_ACCOUNT_GROUP;
 	}
 
-	@Override
-	public boolean isVisible(
-		User user, AccountGroupDisplay accountGroupDisplay) {
-
-		if (accountGroupDisplay.getAccountGroupId() ==
-				AccountConstants.ACCOUNT_ENTRY_ID_DEFAULT) {
-
-			return portalPermission.contains(
-				PermissionCheckerFactoryUtil.create(user),
-				AccountActionKeys.ADD_ACCOUNT_GROUP);
-		}
-
-		return AccountGroupPermission.contains(
-			PermissionCheckerFactoryUtil.create(user),
-			accountGroupDisplay.getAccountGroupId(), ActionKeys.UPDATE);
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/account_groups_admin/account_group/details.jsp");
-	}
-
 	@Reference
-	protected JSPRenderer jspRenderer;
-
-	@Reference
-	protected PortalPermission portalPermission;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
